+++ src/action/interact/confirm/pay.ts
@@ -10,6 +10,10 @@ export interface IPayAssetTransaction {
      * 決済取引番号
      */
     transactionNumber: string;
+    // 確定時の決済方法区分指定(2023-08-29~)
+    object?: {
+        paymentMethod?: { identifier?: string };
+    };
 }
 export type IObject = IPayAssetTransaction[];
 export type IResult = any;