+++ src/orderStatus.ts
@@ -4,10 +4,10 @@
 export enum OrderStatus {
     OrderCancelled = 'OrderCancelled',
     OrderDelivered = 'OrderDelivered',
-    OrderInTransit = 'OrderInTransit',
+    // OrderInTransit = 'OrderInTransit',
     OrderPaymentDue = 'OrderPaymentDue',
-    OrderPickupAvailable = 'OrderPickupAvailable',
-    OrderProblem = 'OrderProblem',
+    // OrderPickupAvailable = 'OrderPickupAvailable',
+    // OrderProblem = 'OrderProblem',
     OrderProcessing = 'OrderProcessing',
     OrderReturned = 'OrderReturned'
 }